rename Registration folder to Signup
